# frontend/openapi.yml
openapi: 3.0.3
info:
  title: CSRS Report API
  version: 1.0.0
  description: API metier versionnee pour l'interface React CSRS Report.
paths:
...
  /api/v1/proposals/:
    get:
      operationId: proposal_list
      tags:
      - proposals
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    post:
      operationId: proposal_create
      tags:
      - proposals
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProposalCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProposalCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/proposals/{id}/:
    get:
      operationId: proposal_detail
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - proposals
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    patch:
      operationId: proposal_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - proposals
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProposalUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProposalUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProposalUpdate'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/proposals/{id}/decision/:
    post:
      operationId: proposal_decision
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - proposals
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalDecision'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProposalDecision'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProposalDecision'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/proposals/{id}/resubmit/:
    post:
      operationId: proposal_resubmit
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - proposals
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalResubmit'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProposalResubmit'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProposalResubmit'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/team/:
    get:
      operationId: team_overview
      parameters:
      - in: query
        name: month
        schema:
          type: string
      - in: query
        name: week
        schema:
          type: string
          format: date
      tags:
      - team
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/team/{id}/:
    get:
      operationId: team_employee_detail
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: query
        name: month
        schema:
          type: string
      - in: query
        name: week
        schema:
          type: string
          format: date
      tags:
      - team
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
components:
  schemas:
...
    DecisionEnum:
      enum:
      - accept
      - reject
      type: string
      description: |-
        * `accept` - accept
        * `reject` - reject
...
    PatchedProposalUpdate:
      type: object
      description: Shared, server-validated assignment schedule.
      properties:
        start_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        estimated_work_days:
          type: string
          format: decimal
          pattern: ^-?\d{0,6}(?:\.\d{0,1})?$
        title:
          type: string
          maxLength: 180
        description:
          type: string
        action_id:
          type: integer
          minimum: 1
          nullable: true
        calendar_id:
          type: integer
          minimum: 1
        revision:
          type: integer
          minimum: 1
...
    ProposalCreate:
      type: object
      description: Shared, server-validated assignment schedule.
      properties:
        start_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        estimated_work_days:
          type: string
          format: decimal
          pattern: ^-?\d{0,6}(?:\.\d{0,1})?$
        title:
          type: string
          maxLength: 180
        description:
          type: string
        action_id:
          type: integer
          minimum: 1
          nullable: true
        calendar_id:
          type: integer
          minimum: 1
      required:
      - description
      - due_date
      - estimated_work_days
      - start_date
      - title
    ProposalDecision:
      type: object
      properties:
        revision:
          type: integer
          minimum: 1
        decision:
          $ref: '#/components/schemas/DecisionEnum'
        reason:
          type: string
          default: ''
      required:
      - decision
      - revision
    ProposalResubmit:
      type: object
      properties:
        revision:
          type: integer
          minimum: 1
      required:
      - revision
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid